pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "k8s-cni-migration-tool"
version = "0.1.0"
description = "A tool to facilitate migration from various Kubernetes CNI solutions to Cilium"
authors = ["CNI Migration Tool Contributors <maintainers@example.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/cmcconnell1/k8s-cni-migration-tool"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
]
packages = [{include = "k8s_cni_migration_tool"}]

[tool.poetry.dependencies]
python = ">=3.8"
kubernetes = "^26.1.0"
pyyaml = "^6.0"
click = "^8.1.3"
rich = "^13.3.5"
jinja2 = "^3.1.2"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"

[tool.poetry.scripts]
cni-migration = "k8s_cni_migration_tool.cli:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
